+++ solver.c
@@ -71,10 +71,10 @@ void solverGPU_HYB(const unsigned int dimension,
 		const unsigned int *row_idx,  const unsigned int* I, const unsigned int *J, const double *V, 
 		const unsigned int totalNumPrecond, const unsigned int *numInRowL, unsigned int maxRowNumPrecond,
 		const unsigned int *row_idxL,  
-		const unsigned int *I_precond, const unsigned int *J_precond, const float* V_precond, 
+		const unsigned int *I_precond, const unsigned int *J_precond, const double* V_precond, 
 		const unsigned int totalNumPrecondP, const unsigned int *numInRowLP, unsigned int maxRowNumPrecondP,
 		const unsigned int *row_idxLP,  
-		const unsigned int *I_precondP, const unsigned int *J_precondP, const double *V_precondP, 
+		const unsigned int *I_precondP, const unsigned int *J_precondP, const double*V_precondP, 
 		const double *vector_in, double *vector_out,  
 		const unsigned int MAXIter, unsigned int *realIter,  const cb_s cb,
 		const unsigned int partition_size, const unsigned int* part_boundary)
@@ -398,7 +398,7 @@ void solverPrecondCPU(const unsigned int procNum, const unsigned int dimension,
 		const unsigned int totalNum, const unsigned int *row_idx, const unsigned int *J, 
 		const double *V, const unsigned int totalNumPrecond, const unsigned int *row_idxL, 
 		const unsigned int *J_precond, const double *V_precond, const unsigned int totalNumPrecondP,
-		const unsigned int *row_idxLP, const unsigned int *J_precondP, const float* V_precondP, 
+		const unsigned int *row_idxLP, const unsigned int *J_precondP, const double* V_precondP, 
 		const double *vector_in, double *vector_out, const unsigned int MAXIter, unsigned int *realIter){
 
 	size_t size0 =dimension*sizeof(double);
